Set up project to use pyenv and virtualenv
This allows us to manage all the repo dependencies locally without
affecting the system dependencies.

It is less than ideal because pyenv sets the version to the actual
virtualenv meaning that we can't specify the Python version. We can only
document it in the README for now until we or someone finds a way to do
that.