Run provisioning in Travis CI
The cost-effective solution we came with to have continuous integration
for our provisioning scripts works as follows:

* Executes Ansible using a local connection, modifying the same CI container
it's running on.
* Because of that, it removes any installed PostgreSQL versions. This
way, we ensure our own database role works.

Also, we had to manually remove PostgreSQL cluster directories.

The `--purge` apt's argument doesn't seem to do it. I tried on a LXC and
running `sudo apt remove --purge postgresql-9.4` opens a ncurses-like UI
to confirm the actual purge of cluster directories. I guess this is what
we can't do when executing it in CI.

Finally, we add an argument to create pg_hba.conf file if it doesn't
exist. It was necessary to make the job succeed.