Install the Hstore PostgreSQL extension
We do it with Ansible with the `postgres` user so we don't need to add
`SUPERUSER` permissions to the app's DB user. Something we don't like.

Then, when deploying the app the extensions listed in `db/schema.rb` are
already in place and the `rake db:migrate` or `rake db:setup` succeeds.