Provision a dedicated CI server instead of Travis
This simply runs the provisioning playbook against a new small VPS we
purchased. This way we can check if our changes are ok in a much more
reliable way than we did so far where the target of the provisioning was
the Travis machine itself. This has become to hackish and makes it hard
to set up ElasticSearch.

For this to work, we need to encrypt and store a new private key while
the public one needs to be uploaded to this server running
`ansible-playbook playbooks/sys_admins.yml --limit=ci_server -u root`
from your local machine first. Otherwise, the `travis` user can't log
into the machine.
https://blog.martignoni.net/2019/03/deploy-your-hugo-site/ is the
article that served as inspiration.

This new key-pair gives granular control over the provisioning from CI.
If something bad happens we can create a new one.

Unfortunately, we need the Travis CLI to set all this up which is
distributed as a Ruby gem. You can find all the details in the
[README](https://github.com/coopdevs/timeoverflow-provisioning#CI).

Something will need fixing soon is the error `Timeout when waiting for
localhost:9200` we get very often as a result of the task
`[elastic.elasticsearch : Wait for elasticsearch to startup]`. You can
get going with a retry but that misses the point of CI.

Also, this creates a sysadmin user for Travis in the target machine.
While this does the trick we might not want such level of permissions
for that user in the future. Better to stick to the lowest level as
possible.