Skip provisioning CI when building master
When building from the master branch we want to provision staging but
provisioning the CI server is totally unnecessary and, as it is, it
could create concurrency problems if another PR is being built. It
happened to me that two different CI builds were fighting for the dpkg
lock and one failed.