Provision production from CI when staging passed
This will be useful now that we want to setup a new production server
quickly. Every PR we merged will be provisioned there only if it
succeeds in staging so we don't forget about keeping it up-to-date as we
move forward.

This will be removed once that machine is fully set up right before we
promote it to become the actual production server.

Note this points to a `production` host that doesn't exist yet. I'll
open a PR to update the inventory which will kick off the process.