Add script used to rebuild a Hetzner server
I had this untracked file for a while and it annoys me in every commit.
It's the file I used a few times to rebuild a server when setting up
staging. It help when the server was doing nasty things as well as
recovering from a broken provisioning.

However, since https://github.com/coopdevs/donalo/pull/12 we now know
there's a better way to manage the state of a Hetzner server.