Add rollbar_access_token to all hosts secrets
We opened a Rollbar account 5 years ago to track errors in production
which was manually set as and ENV var in production but never got added
to our Ansible playbooks with which we managed staging so far. Let's use
it with all environments then.

Although it was added separately to each host_vars' secrets.yml they all
share the same token with `post_server_item` scope. It could likely
a group_vars' setting to reflect this but we can improve that later on.