Add unicorn stop and db drop to restore a backup
This steps provided necessary to properly restore a production backup in
next. First, we need to kill the existing database connections and then,
drop the existing database so we don't cause data inconsistencies
current and new records.

We did so with a user prompt to double check your pointing to the right
database.