Dry paths with app_path and current_path vars
This makes the code a bit easier to understand as well. You know at
a glance which things refer to the current release.

Note that we also removed the creation of `/var/www`. According to the
docs there is no need for that middle step

> If directory, all intermediate subdirectories will be created if they do not exist.

We can create the whole app_path in one shot.