Fix EnvironmentFile systemd directive
We broke the path to the user defaults in
https://github.com/coopdevs/timeoverflow-provisioning/pull/150.

We also discovered the meaning of that `-`:

> The "-" on the EnvironmentFile= line ensures that no error messages is
generated if the environment file does not exist. Since many of these
files were optional in sysvinit, you should include the "-" when using
this directive.

sources:
https://fedoraproject.org/wiki/Packaging:Systemd#EnvironmentFiles_and_support_for_.2Fetc.2Fsysconfig_files, https://www.freedesktop.org/software/systemd/man/systemd.exec.html#EnvironmentFile=

We don't want that dash. If the file is not found we want to know, not
to fail silently. The env vars it contains are nowhere else.

And it turns out this also fixes Monday's digest email.