Backup assets only
Assets are what could go lost if there's a serious issue with the disk.
The app code is stored in the Github repo so to recover from a failure
we would deploy again, restore the assets from the backup and the DB
from the dump.

Backing up anything else beyond public/ is becoming tricky. Log files change
while issuing the tar command (it doesn't like that) and I'm struggling
trying to exclude the log/ and tmp/ directories. This way things are way
simpler.