Boot Unicorn using bundle exec instead of binstub
We removed bin/unicorn and bin/unicorn_rails in
https://github.com/coopdevs/timeoverflow/commit/29c8ec03fc5aef28fb59ca2e24ba49345a4c368d,
which was used to start the systemd unit. Those binstubs don't add much
and we can simply use `bundle exec`. We also make use of the
`WorkingDirectory` directive and avoid full paths for better
readability.

Also, it looks like we introduced Unicorn back in
https://github.com/coopdevs/timeoverflow/commit/b8fc45aad80fbda06d2b83dd70da15e776c6bdb2
but never really added the gem
https://github.com/samuelkadolph/unicorn-rails for which we did add
a binstub. Perhaps the change in the Gemfile was never committed?